refactor: Rename database file
