+++ src/auth.ts
@@ -2,7 +2,7 @@ import { PrismaAdapter } from "@auth/prisma-adapter";
 import NextAuth from "next-auth";
 import { authConfig } from "./auth.config";
 import { onCreateUser } from "./functions/user/onCreateUser";
-import { prisma } from "./server/db/client";
+import { prisma } from "./server/db";
 
 export const { handlers, auth, signIn, signOut } = NextAuth({
   ...authConfig,
