+++ src/dal/mealPlans/getMealPlans.ts
@@ -1,4 +1,4 @@
-import { prisma } from "@/server/db/client";
+import { prisma } from "@/server/db";
 
 /**
  * Gets all mealplans for this user
