+++ src/dal/mealPlans/submitMealEntry.ts
@@ -1,4 +1,4 @@
-import { prisma } from "@/server/db/client";
+import { prisma } from "@/server/db";
 import { type MealEntry } from "@prisma/client";
 
 type Props = {
