+++ src/dal/user/getInvitation.ts
@@ -1,4 +1,4 @@
-import { prisma } from "@/server/db/client";
+import { prisma } from "@/server/db";
 import type { MealPlan, MealPlanInvite, User } from "@prisma/client";
 
 /**
