+++ src/dal/user/redeemMealPlanInvitation.ts
@@ -1,4 +1,4 @@
-import { prisma } from "@/server/db/client";
+import { prisma } from "@/server/db";
 import type { MealPlanInvite } from "@prisma/client";
 
 /**
